requirements.txt
Django>=3.1.1,<3.2
django-cleanup>=5.1.0,<5.2
djangorestframework>=3.11.1,<3.12
drf-yasg>=1.20.0,<1.21
psycopg2-binary>=2.8.5,<2.9
stripe==2.55.0
